% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band.R
\name{band}
\alias{band}
\title{Extract Regions by Date}
\usage{
band(
  imageCol,
  geeFC = NULL,
  scale,
  band = NULL,
  temporal = "yearly",
  temporal_stat = "median",
  lazy = FALSE,
  reducer = "median",
  variable = NULL,
  ggplot = FALSE,
  save.plot = F,
  user_geom = NULL,
  startDate = NULL,
  endDate = NULL,
  c.low = NULL,
  c.high = NULL
)
}
\arguments{
\item{imageCol}{A previously created get_* object or ee.ImageCollection}

\item{geeFC}{A known GEE FeatureCollection or asset, e.g. "USGS/WBD/2017/HUC12"}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 800 (m) default.}

\item{band}{A \code{character} indicating what bands/type to use when you have more than one.}

\item{temporal}{A \code{character} indicating what temporal filter to use on the collection, e.g. 'yearly' (default), 'monthly', 'year_month', 'all'.}

\item{temporal_stat}{A \code{character} indicating what to reduce the ImageCollection when using temporal filtering, e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'stdDev', 'first'.}

\item{lazy}{\code{logical} whether to run a 'sequential' future in the background or not.}

\item{reducer}{A \code{character} 'median' (default) for reducing.}

\item{variable}{\code{character} indicating what to facet ggplot by. Need to know ahead of time.}

\item{ggplot}{\code{logical} TRUE/FALSE. Whether to print side-effect ggplot. See details.}

\item{save.plot}{\code{logical} TRUE/FALSE. Whether to save the plot in a list.}

\item{user_geom}{\code{logical} TRUE/FALSE.}

\item{startDate}{\code{character} format date, e.g. "2018-10-23". NULL (default)}

\item{endDate}{\code{character} format date, e.g. "2018-10-23". NULL (default)}

\item{c.low}{\code{numeric} lower month value for calendar range. NULL (default)}

\item{c.high}{\code{numeric} higher month value for calendar range. NULL (default)}
}
\value{
A \code{data.frame} and a side-effect plot (if ggplot = TRUE); unless using \code{save.plot} then a list with \code{data.frame} and ggplot.
}
\description{
This function allows the user to pass a previously created get_*() object to get
a time series from a collection, e.g. banding/toBands. This uses the \link[rgee]{ee_extract} function from rgee.
}
\note{
If using a get_\emph{() object band() doesn't use the image generated. It only uses the ImageCollection.
Additional arguments to pass if using a ImageCollection without get_}()'s are the startDate, endDate, c.low and c.high arguments.
Faster with points or centroids of polygons.
If lazy is TRUE, the function will be run in the background.
}
\examples{
\dontrun{

# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

ld8 <- get_landsat(huc, method = 'ld8', startDate = '2014-01-01',
                  endDate = '2018-12-31', c.low = 6, c.high = 11)

# without plotting save to object
ld8_ts <- ld8 \%>\% band(scale = 30, band = 'NDVI')

# with plotting as side-effect
ld8 \%>\% band(scale = 30, band = 'NDVI', ggplot = TRUE, variable = 'name')

# save both plot and data
ld8_ts <- ld8 \%>\% band(scale = 30, band = 'NDVI',
                       ggplot = TRUE, variable = 'name',
                       save.plot = TRUE)

}
}
