% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPP.R
\name{band_NPP}
\alias{band_NPP}
\title{Extract Regions by Date}
\usage{
band_NPP(
  aoi,
  wb,
  id,
  startDate = "1986-01-01",
  endDate = "2019-01-01",
  scale = 30,
  crs = 4326,
  plot = TRUE,
  save.plot = "FALSE"
)
}
\arguments{
\item{aoi}{A sf object.}

\item{wb}{A \code{character} of HUC id, e.g. HUC02-HUC12.}

\item{id}{A \code{vector} identifying each region. \code{optional} if using \code{wb}}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 10 (m) default.}

\item{crs}{\code{numeric} crs code}

\item{plot}{\code{logical} TRUE/FALSE.}

\item{save.plot}{\code{logical} TRUE/FALSE whether to save plot and data.frame}
}
\value{
A \code{data.frame} and a side-effect plot; unless using \code{save.plot}.
}
\description{
This function is similar to \link[exploreRGEE]{rr_NPP} but instead of reducing the region
for entire date range this function keeps the dates selected ('year' and 'months') and returns the median of each month.
This function is good if your wanting a time series within a region(s).
}
