% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM.R
\name{band_PRISM}
\alias{band_PRISM}
\title{Extract Regions by Date}
\usage{
band_PRISM(
  aoi,
  id,
  wb,
  method = "AN81m",
  param = "ppt",
  startDate = "1981-01-01",
  endDate = "2010-01-01",
  c.low = 1,
  c.high = 12,
  plot = TRUE,
  save.plot = "FALSE",
  scale = 800,
  crs = 4326
)
}
\arguments{
\item{aoi}{A sf object.}

\item{id}{A \code{vector} indentifying each region.}

\item{wb}{A \code{character} of HUC id, e.g. HUC02-HUC12.}

\item{method}{\code{character}. 'AN81m'(default), 'AN81d'.}

\item{param}{\code{character}. 'ppt' (default), tmin, tmax, tmean, tdmean (deg C); vpdmin, vpdmax (hPa)}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}

\item{plot}{\code{logical} TRUE/FALSE.}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 800 (m) default.}

\item{crs}{\code{numeric} crs value}
}
\value{
A \code{data.frame} and a side-effect plot; unless using \code{save.plot}.
}
\description{
This function is similar to \link[exploreRGEE]{rr_PRISM} but instead of reducing the region
post date range (only returning one statistic value) this function keeps the dates selected, e.g. 'months' and returns the median of each month. It uses only the 'AN81m' or 'AN81d' image
collections for processing. This function is good if your wanting a time series within a region.
}
\note{
Might change this function in the future by using map() server-side to speed up computations. Better with points or centroids of polygons.
}
