% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent2.R
\name{band_Sent2}
\alias{band_Sent2}
\title{Extract Regions by Date}
\usage{
band_Sent2(
  aoi,
  wb,
  id,
  param = "NDVI",
  startDate = "2019-04-01",
  endDate = "2020-10-01",
  c.low = 1,
  c.high = 12,
  scale = 250,
  plot = TRUE,
  save.plot = "FALSE"
)
}
\arguments{
\item{aoi}{A sf object.}

\item{wb}{A \code{character} of HUC id, e.g. HUC02-HUC12.}

\item{id}{A \code{vector} identifying each region. \code{optional} if using \code{wb}}

\item{param}{A \code{character} indicating what band to visualize, e.g. 'Blue', 'Green', 'Red', 'NIR', 'NDVI', 'NDWI'.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 250 (m) default.}

\item{plot}{\code{logical} TRUE/FALSE.}

\item{save.plot}{\code{logical} TRUE/FALSE whether to save plot and data.frame}
}
\value{
A \code{data.frame} and a side-effect plot; unless using \code{save.plot}.
}
\description{
This function is similar to \link[exploreRGEE]{rr_Landsat} but instead of reducing the region
for entire date range this function keeps the dates selected ('year' and 'months') and returns the median of each month.
This function is good if your wanting a time series within a region(s).
}
