% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ee_viz}
\alias{ee_viz}
\alias{ee_viz.ee.image.Image}
\alias{ee_viz.exploreList}
\title{Viz-ualize GEE Image/ImageCollections}
\usage{
ee_viz(x, ...)

\method{ee_viz}{ee.image.Image}(
  x,
  scale = 1000,
  geom = NULL,
  band = NULL,
  palette = hcl.colors(11, "RdBu"),
  range = NULL,
  gamma = NULL,
  opacity = NULL,
  ...
)

\method{ee_viz}{exploreList}(
  x,
  scale = 1000,
  band = NULL,
  palette = hcl.colors(11, "RdBu"),
  range = NULL,
  gamma = NULL,
  opacity = NULL,
  ...
)
}
\arguments{
\item{x}{A previously created get_* object or ee.image.Image}

\item{...}{extra args to pass on}

\item{scale}{A \code{numeric} value indicating scale in meters. 250 (default)}

\item{geom}{A sf object to use as 'aoi/geom' for an 'ee.image.Image'.}

\item{band}{A \code{character} indicating what bands/type to visualize.
Can select more than one, e.g. c('Red', 'Green', 'Blue').}

\item{palette}{\code{vector} of a color palette.}

\item{range}{\code{vector} indicating a numerical range to set min/max visualization.}

\item{gamma}{\code{numeric} gamma correction factor.}

\item{opacity}{\code{numeric} transparent display value.}
}
\value{
A leaflet map.
}
\description{
This function allows users to quickly view a previously created get_*() object or ee.image.Image.
}
\note{
This function uses a scale argument which is used to generate a min and max value for viewing.
Because this uses getInfo(), it can take a while
depending on the scale. Since this is used for viewing, I would suggest to go bigger on the scale. If a user selects more than one band,
the 'up-to' three bands will be overlayed like earth engine.
}
\examples{
\dontrun{

# Load Libraries

library(rgee)
ee_Initialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

ld8 <- get_landsat(huc, method = 'ld8', startDate = '2014-01-01',
                  endDate = '2018-12-31', c.low = 6, c.high = 11)

ld8 \%>\% ee_viz(scale = 30, band = 'NDVI', palette = 'RdYlGn')

# or without scale, when min and max are used scale is irrelevant.

ld8 \%>\% ee_viz(min = 0, max = 1, band = 'NDVI', palette = 'RdYlGn')

}


}
