% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{extract_time}
\alias{extract_time}
\title{extract_time}
\usage{
extract_time(imageCol, reg, reducer, scale, ...)
}
\description{
Function used to get extraction of filtered IC and then returns a nice tidy
data.frame with time = date and .names as the band name, e.g. B1_mean, B1_median, etc.
}
