% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_any}
\alias{get_any}
\title{Get Earth Engine Products}
\usage{
get_any(
  aoi,
  i_type = "ImageCollection",
  method,
  param = NULL,
  stat = "median",
  startDate = NULL,
  endDate = NULL,
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  c.low = NULL,
  c.high = NULL
)
}
\arguments{
\item{aoi}{A sf object indicating the extent of the geom.}

\item{i_type}{A \code{character} indicating what type of image, e.g. 'ImageCollection' or 'Image'.}

\item{method}{A \code{character} indicating what imageCollection to use, e.g. "UMD/hansen/global_forest_change_2019_v1_7".}

\item{param}{A \code{character} indicating what band to select.}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by, e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'stdDev', 'first'.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{mask}{\code{logical} whether to mask out certain ranges}

\item{m.low}{\code{numeric} low value for mask, e.g. greater than 'm.low'}

\item{m.high}{\code{numeric} high value for mask, e.g. less than 'm.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}
}
\value{
A list of Earth Engine Objects and arguments.
}
\description{
This function allows the user to provide a earth engine image/imageCollection \code{character} string which will help with simple processing.
}
\note{
Use how you would normally call a GEE session online but with method as your collection snippet.
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

# get any Image or ImageCollection

forest_cover_loss <- get_any(huc, i_type = "Image",
                             method = 'UMD/hansen/global_forest_change_2019_v1_7',
                             param = 'lossyear')
}

}
