% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{get_diff}
\alias{get_diff}
\title{Title}
\usage{
get_diff(
  aoi,
  method = "LANDSAT",
  param = "NDVI",
  startDate = "2000-01-01",
  endDate = "2010-12-30",
  startDate2 = "2010-01-01",
  endDate2 = "2020-10-30",
  c.low = 1,
  c.high = 12,
  scale = 250
)
}
\arguments{
\item{aoi}{A sf object}

\item{method}{A \code{character} indicating what method to use, e.g. 'LANDSAT', 'PRISM', 'MODIS', 'SAR2'}

\item{param}{A \code{character} indicating what band to visualize. Check other functions to find out.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{startDate2}{\code{character} format date, e.g. "1999-10-23". A second start date to use as the subtraction window.}

\item{endDate2}{\code{character} format date, e.g. "1999-10-23". A second end date to use as the subtraction window.}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 250 (m) default.}
}
\value{
A leaflet map.
}
\description{
Title
}
