% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_linear}
\alias{get_linear}
\title{Get Image Difference}
\usage{
get_linear(
  data,
  method = "lfit",
  band = NULL,
  stat = "median",
  temporal = "yearly"
)
}
\arguments{
\item{data}{A previously create get_* object.}

\item{band}{A \code{character} indicating what bands/type to use when you have more than one. Can \strong{only} select one, e.g. 'NDVI'.}

\item{stat}{A \code{character} indicating what to reduce the imageCollection when using temporal filtering, e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'stdDev', 'first'.}

\item{temporal}{A \code{character} indicating what temporal filter to use on the collection, e.g. 'yearly' (default), 'year_month', 'all'.}
}
\value{
A list of Earth Engine Objects and arguments. Specifically an ImageCollection with
bands c('scale', 'offset').
}
\description{
This function allows the user to get a linearFit() from an ImageCollection.
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

# get intitial image. Can by any get_*().

ld8 <- get_landsat(huc, method = 'ld8', startDate = '2014-01-01',
                  endDate = '2018-12-31', c.low = 6, c.high = 11)

# now subtract from new one

lr <- ld8 \%>\% get_linear(band = 'NDVI', temporal = 'yearly')

lr \%>\% viz(band = 'scale')
}
}
