% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_met}
\alias{get_met}
\title{Get Meteorological Products}
\usage{
get_met(
  aoi,
  method = "Norm81m",
  param = NULL,
  stat = "median",
  startDate = "1981-01-01",
  endDate = "2010-01-01",
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  c.low = 1,
  c.high = 12
)
}
\arguments{
\item{aoi}{A sf object indicating the extent of the geom.}

\item{method}{\code{character}. 'Norm81m'(default), see details.}

\item{param}{\code{character}. NULL (default). Use the band names for appropriate dataset method, e.g. PRISM = 'ppt', GRIDMET = 'pr', DAYMET = 'prcp', etc.}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by, e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'stdDev', 'first'.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{mask}{\code{logical} whether to mask out certain ranges}

\item{m.low}{\code{numeric} low value for mask, e.g. greater than 'm.low'}

\item{m.high}{\code{numeric} high value for mask, e.g. less than 'm.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}
}
\value{
A list of Earth Engine Objects and arguments.
}
\description{
This function gets you all the Meteorological data you want in a Earth Engine object. See details for available methods.
}
\details{
The methods currently available (more to come):
\itemize{
\item  \strong{Norm81m}: PRISM 1981-2010; Monthly, annual average.
\item \strong{AN81m}: PRISM Jan 1895 - ongoing; Monthly annual time series.
\item \strong{AN81d}: PRISM Jan 1 1981 - ongoing; Daily time series.
\item \strong{GRIDMET}: 1979-01-01 - ongoing; The Gridded Surface Meteorological dataset provides high spatial resolution (~4-km) daily surface fields of temperature, precipitation, winds, humidity and radiation across the contiguous United States from 1979. \cr
\item  \strong{DAYMET}: 1980-01-01 - 2019-12-31; Daymet V3 provides gridded estimates of daily weather parameters for United States, Mexico, Canada, Hawaii, and Puerto Rico. It is derived from selected meteorological station data and various supporting data sources. \cr
\item \strong{TRMMh}: 1998-01-01 - 2020-12-31; The Tropical Rainfall Measuring Mission (TRMM) is a joint mission between NASA and the Japan Aerospace Exploration Agency (JAXA) designed to monitor and study tropical rainfall. The 34B2 product contains a gridded, TRMM-adjusted, merged infrared precipitation (mm/hr) and RMS precipitation-error estimate, with a 3-hour temporal resolution and a 0.25 degree spatial resolution. \cr
\item \strong{TRMMm}: 1998-01-01 - 2019-12-01; This collection is no longer being updated. See IMERG monthly. This dataset algorithmically merges microwave data from multiple satellites, including SSMI, SSMIS, MHS, AMSU-B and AMSR-E, each inter-calibrated to the TRMM Combined Instrument.
\item \strong{TERRACLIMATE}: 1958-01-01 - 2019-12-01; TerraClimate is a dataset of monthly climate and climatic water balance for global terrestrial surfaces.
}
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

prism_monthly <- get_met(huc, method = 'AN81m', startDate = '2014-01-01',
                  endDate = '2018-12-31', c.low = 6, c.high = 11)


}
}
