% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_nex}
\alias{get_nex}
\title{NASA Earth Exchange Products}
\usage{
get_nex(
  aoi,
  method = "ensemble",
  scenario = "rcp85",
  model = "ACCESS1-0",
  param = NULL,
  stat = "median",
  startDate = NULL,
  endDate = NULL,
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  c.low = NULL,
  c.high = NULL
)
}
\arguments{
\item{aoi}{A sf object indicating the extent of the geom.}

\item{method}{A \code{character} indicating what imageCollection to use, e.g. "ensemble" is equal to = "NASA/NEX-DCP30_ENSEMBLE_STATS".}

\item{scenario}{A \code{character} indicating what Representative Concentration Pathways (RCPs) to use, e.g. 'rcp85' (default).}

\item{model}{A \code{character} indicating what model to use (only for 'dcp' and 'gddp'). 'ACCESS1-0' (default).}

\item{param}{A \code{character} indicating what band to select.}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by, e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'stdDev', 'first'.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{mask}{\code{logical} whether to mask out certain ranges}

\item{m.low}{\code{numeric} low value for mask, e.g. greater than 'm.low'}

\item{m.high}{\code{numeric} high value for mask, e.g. less than 'm.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}
}
\value{
A list of Earth Engine Objects and arguments.
}
\description{
This function allows the user to get NASA Earth Exchange products.
}
\details{
The \code{method}, \code{scenario} and \code{model} currently available:
\itemize{
\item  \strong{method}: ensemble, dcp, gddp.
\item  \strong{scenario}: 'historical', 'rcp26', 'rcp45', 'rcp60', 'rcp85'
\item \strong{model}: ACCESS1-0', 'bcc-csm1-1', 'bcc-csm1-1-m', 'BNU-ESM', 'CanESM2', 'CCSM4', 'CESM1-BGC', 'CESM1-CAM5', 'CMCC-CM', 'CNRM-CM5', 'CSIRO-Mk3-6-0', 'FGOALS-g2', 'FIO-ESM', 'GFDL-CM3', 'GFDL-ESM2G', 'GFDL-ESM2M', 'GISS-E2-H-CC', 'GISS-E2-R', 'GISS-E2-R-CC', 'HadGEM2-AO', 'HadGEM2-CC', 'HadGEM2-ES', 'inmcm4', 'IPSL-CM5A-LR', 'IPSL-CM5A-MR', 'IPSL-CM5B-LR', 'MIROC5', 'MIROC-ESM', 'MIROC-ESM-CHEM', 'MPI-ESM-LR', 'MPI-ESM-MR', 'MRI-CGCM3', 'NorESM1-M'
}
}
\note{
\code{scenario} only uses \code{dcp} and \code{gddp} so disregard when using \code{ensemble} and be aware when using the others. In addition, it is up to the users to understand what model to use with
what scenario! Some scenarios and models don't jive so do your research.
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

# get any Image or ImageCollection

forest_cover_loss <- get_any(huc, method = 'ensemble',
                             scenario = 'rcp85',
                             model = 'ACCESS1-0',
                             method = 'NASA/NEX-DCP30_ENSEMBLE_STATS',
                             param = 'lossyear')
}

}
