% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_npp}
\alias{get_npp}
\title{Get Net Annual NPP}
\usage{
get_npp(
  aoi,
  method = "cloud_mask",
  param = "annualNPP",
  stat = "median",
  startDate = "1986-01-01",
  endDate = "2019-01-01",
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  c.low = 1,
  c.high = 12
)
}
\arguments{
\item{aoi}{A sf object indicating the extent of the geom.}

\item{method}{A \code{character} indicating what method to use, e.g. 'cloud_mask', 'raw'.}

\item{param}{A \code{character} indicating what band to visualize, e.g. 'annualNPP'.}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by, e.g. 'mean', 'median', 'max', 'min', 'sum', 'stdDev', 'first'.}

\item{startDate}{\code{character} format date, e.g. "2018-10-23"}

\item{endDate}{\code{character} format date, e.g. "2018-10-23"}

\item{mask}{\code{logical} whether to mask out certain ranges}

\item{m.low}{\code{numeric} low value for mask, e.g. greater than 'm.low'}

\item{m.high}{\code{numeric} high value for mask, e.g. less than 'm.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}
}
\value{
A list of Earth Engine Objects and arguments.
}
\description{
This function uses the 30-m resolution image collection of Landsat Net Primary Productivity
to get selected time frames. There's really only one product here so 'cloud_mask' is a method instead of a logical.
}
\note{
The Landsat Net Primary Production (NPP) CONUS dataset
estimates NPP using Landsat Surface Reflectance for CONUS.
NPP is the amount of carbon captured by plants in an ecosystem,
after accounting for losses due to respiration.
NPP is calculated using the MOD17 algorithm (see MOD17 User Guide)
with Landsat Surface Reflectance, gridMET, and the National Land Cover Database. Some composites were masked out because
of missing data, high cloud contamination, and/or erroneous pixels which didn't meet the gap-filling method.
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

npp <- get_npp(huc, method = 'cloud_mask', param = 'annualNPP', startDate = '2014-01-01',
                  endDate = '2018-12-31')


}
}
