% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_sent2}
\alias{get_sent2}
\title{Get Sentinel-2 Products}
\usage{
get_sent2(
  aoi,
  method = "S2_1C",
  param = NULL,
  stat = "median",
  cloud_mask = TRUE,
  startDate = "2015-04-01",
  endDate = "2020-01-01",
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  c.low = 1,
  c.high = 12
)
}
\arguments{
\item{aoi}{A sf object indicating the extent of the geom.}

\item{method}{A \code{character} indicating what method to use, e.g. 'S2_1C' (default) or 'S2_2A'.}

\item{param}{A \code{character} indicating what band to visualize, e.g. 'Blue', 'Green', 'Red', 'NIR', 'NDVI', 'NDWI', 'NBR', etc.}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by, e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'stdDev', 'first'.}

\item{cloud_mask}{\code{logical} whether to mask out certain cloud artifacts. TRUE (default).}

\item{startDate}{\code{character} format date, e.g. "2018-10-23"}

\item{endDate}{\code{character} format date, e.g. "2018-10-23"}

\item{mask}{\code{logical} whether to mask out certain ranges}

\item{m.low}{\code{numeric} low value for mask, e.g. greater than 'm.low'}

\item{m.high}{\code{numeric} high value for mask, e.g. less than 'm.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}
}
\value{
A list of Earth Engine Objects and arguments.
}
\description{
This function uses the Sentinel-2 missions.
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

sent2_2A_SCL <- get_sent2(huc, method = 'S2_2A', param = 'SCL', startDate = '2018-01-01',
                  endDate = '2020-12-31', c.low = 6, c.high = 11)


}
}
