% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_terrain}
\alias{get_terrain}
\title{Get Terrain Products}
\usage{
get_terrain(
  aoi,
  method = "NED",
  param = "slope",
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  az = 270,
  el = 45
)
}
\arguments{
\item{aoi}{A sf object indicating the extent of the geom.}

\item{method}{\code{character} indicating what method to use, e.g. 'NED', 'SRTM'.}

\item{param}{\code{character} indicating terrain type, e.g. 'dem', 'FA', 'TWI', 'TRI', 'TPI', 'aspect', 'slope', 'cos', 'sin', 'hillshade', 'complete'.}

\item{mask}{\code{logical} whether to mask out certain ranges}

\item{m.low}{\code{numeric} low value for mask, e.g. greater than 'm.low'}

\item{m.high}{\code{numeric} high value for mask, e.g. less than 'm.high'}

\item{az}{\code{numeric} The illumination azimuth in degrees from north.}

\item{el}{\code{numeric} The illumination elevation in degrees.}
}
\value{
A list of Earth Engine Objects and arguments.
}
\description{
This function takes the USGS NED (National Elevation Dataset)
or SRTM (Shuttle Radar Topography Mission) and gets a handful of terrain indices. This is good for
downloaded areas for further analysis or passing on for some quick stats.
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

# get terrain

twi <- get_terrain(huc, method = 'NED', param = 'TWI')
}
}
