% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_video.R
\name{get_Video}
\alias{get_Video}
\title{Visualise NDVI}
\usage{
get_Video(
  aoi,
  collection = "MOD13",
  param = NULL,
  startDate = "2018-01-01",
  endDate = "2019-01-01"
)
}
\arguments{
\item{endDate}{}
}
\value{
A gif video.
}
\description{
Visualise NDVI
}
