% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huc}
\alias{huc}
\title{Hydrological Unit Codes (12-field), e.g. HUC12.}
\format{
A sf object with 2 rows and 11 variables:
\describe{
\item{name}{watershed name}
\item{geometry}{polygon geometry}
...
}
}
\usage{
huc
}
\description{
A sf object with 2 HUC12 catchments
}
\keyword{datasets}
