% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_landsat.R
\name{maskcloud1}
\alias{maskcloud1}
\title{Function to mask cloud from built-in quality band
information on cloud}
\usage{
maskcloud1(image)
}
\arguments{
\item{image}{}
}
\description{
Function to mask cloud from built-in quality band
information on cloud
}
