% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDVI.R
\name{rr_Landsat}
\alias{rr_Landsat}
\title{Get Landsat Reduced Regions}
\usage{
rr_Landsat(
  aoi,
  wb,
  param = "NDVI",
  startDate = "1984-04-01",
  endDate = "2020-01-01",
  c.low = 1,
  c.high = 12,
  leaflet = TRUE,
  scale = 250,
  crs = 4326
)
}
\arguments{
\item{aoi}{A sf object.}

\item{wb}{A \code{character} of HUC id, e.g. HUC02-HUC12.}

\item{param}{A \code{character} indicating what band to visualize, e.g. 'Blue', 'Green', 'Red', 'NIR', 'NDVI', 'NDWI'.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}

\item{leaflet}{\code{logical} indicating to view leaflet map. default (TRUE)}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 250 (m) default.}

\item{crs}{\code{numeric}}
}
\value{
An sf object and leaflet map.
}
\description{
This function uses landsat and any sf object or USGS Watershed Boundary Dataset of Basins to
reduce regions to commmon statistics: mean, max, min, median, stdDev and sum for selected date ranges.
}
