% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPP.R
\name{rr_NPP}
\alias{rr_NPP}
\title{Get Landsat Reduced Regions}
\usage{
rr_NPP(
  aoi,
  wb,
  startDate = "1986-01-01",
  endDate = "2019-01-01",
  leaflet = TRUE,
  scale = 250,
  crs = 4326
)
}
\arguments{
\item{aoi}{A sf object.}

\item{wb}{A \code{character} of HUC id, e.g. HUC02-HUC12.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{leaflet}{\code{logical} indicating to view leaflet map. default (TRUE)}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 250 (m) default.}

\item{crs}{\code{numeric}}
}
\value{
An sf object.
}
\description{
This function uses NPP and any sf object or USGS Watershed Boundary Dataset of Basins to
reduce regions to commmon statistics: mean, max, min, median, stdDev and sum for selected date ranges.
}
