% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM.R
\name{rr_PRISMwb}
\alias{rr_PRISMwb}
\title{Get PRISM Reduced Regions}
\usage{
rr_PRISMwb(
  aoi,
  wb,
  method = "Norm81m",
  param = "ppt",
  startDate = "1981-01-01",
  endDate = "2010-01-01",
  c.low = 1,
  c.high = 12,
  crs = 4326,
  scale = 800,
  leaflet = TRUE
)
}
\arguments{
\item{aoi}{A sf object.}

\item{wb}{A \code{character} of HUC id, e.g. HUC02-HUC12.}

\item{method}{\code{character}. 'Norm81m'(default), 'AN81m', 'AN81d'.}

\item{param}{\code{character}. 'ppt' (default), tmin, tmax, tmean, tdmean (deg C); vpdmin, vpdmax (hPa)}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}

\item{crs}{\code{numeric} crs code}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 800 (m) default.}

\item{leaflet}{\code{logical} indicating to view leaflet map. default (TRUE)}
}
\value{
An sf object.
}
\description{
This function uses PRISM and the USGS Watershed Boundary Dataset of Basins to
reduce regions (wb) to commmon statistics, e.g. mean, max, min, median, stdDev and sum.
}
