% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{rr_Terrain}
\alias{rr_Terrain}
\title{Get slope Reduced Regions}
\usage{
rr_Terrain(aoi, wb, leaflet = TRUE, crs = 4326, scale = 30)
}
\arguments{
\item{aoi}{A sf object.}

\item{wb}{A \code{character} of HUC id, e.g. HUC02-HUC12.}

\item{leaflet}{\code{logical} TRUE/FALSE whether to include leaflet map}

\item{crs}{\code{numeric}}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 800 (m) default.}
}
\value{
An sf object and leaflet map.
}
\description{
This function uses USGS NED and any sf object or USGS Watershed Boundary Dataset of Basins to
reduce regions to commmon statistics: mean, max, min, median, stdDev and sum for selected date ranges.
}
