% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{temporal_filter}
\alias{temporal_filter}
\title{Filtering by Temporal Arguments}
\usage{
temporal_filter(temporal, imageCol, temporal_stat)
}
\arguments{
\item{temporal}{time to reduce}

\item{imageCol}{ee ImageCollection}

\item{temporal_stat}{what to reduce time by.}
}
\description{
A helper function to distinguish time reducer
}
