% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vizA.R
\name{viz_A}
\alias{viz_A}
\title{Leaflet basemap}
\usage{
viz_A()
}
\value{
A leaflet map with common basemaps ("Esri.WorldImagery","CartoDB.Positron", "OpenStreetMap",
"CartoDB.DarkMatter",
"OpenTopoMap")
}
\description{
Leaflet basemap
}
