% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_helper}
\alias{viz_helper}
\title{viz_helper}
\usage{
viz_helper(x, geom, band, scale, range, gamma, opacity, palette)
}
\arguments{
\item{x}{image}

\item{geom}{geometry of image}

\item{band}{a band(s) input}

\item{scale}{what to scale the min/max getInfo() by}

\item{range}{min/max vector}

\item{gamma}{gamma value}

\item{opacity}{opacity of image}

\item{palette}{palette to use}
}
\description{
viz_helper
}
