% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDVI.R
\name{viz_Landsat}
\alias{viz_Landsat}
\title{Visualize Landsat}
\usage{
viz_Landsat(
  aoi,
  param = "NDVI",
  startDate = "1986-04-01",
  endDate = "2020-10-30",
  window = FALSE,
  w.low = NULL,
  w.high = NULL,
  c.low = 1,
  c.high = 12,
  scale = 250
)
}
\arguments{
\item{aoi}{A sf object}

\item{param}{A \code{character} indicating what band to visualize, e.g. 'Blue', 'Green', 'Red', 'NIR', 'NDVI', 'NDWI'.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{window}{\code{logical} whether to mask out certain ranges}

\item{w.low}{\code{numeric} low value for mask, e.g. greater than 'w.low'}

\item{w.high}{\code{numeric} high value for mask, e.g. less than 'w.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 250 (m) default.}
}
\value{
A leaflet map.
}
\description{
Visualize Landsat
}
