% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPP.R
\name{viz_NPP}
\alias{viz_NPP}
\title{Visualize Net Annual NPP}
\usage{
viz_NPP(
  aoi,
  startDate = "1986-01-01",
  endDate = "2019-01-01",
  window = FALSE,
  w.low = NULL,
  w.high = NULL,
  scale = 250
)
}
\arguments{
\item{aoi}{A sf object}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{window}{\code{logical} whether to mask out certain ranges}

\item{w.low}{\code{numeric} low value for mask, e.g. greater than 'w.low'}

\item{w.high}{\code{numeric} high value for mask, e.g. less than 'w.high'}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 250 (m) default.}
}
\value{
A leaflet map.
}
\description{
The Landsat Net Primary Production (NPP) CONUS dataset
estimates NPP using Landsat Surface Reflectance for CONUS.
NPP is the amount of carbon captured by plants in an ecosystem,
after accounting for losses due to respiration.
NPP is calculated using the MOD17 algorithm (see MOD17 User Guide)
with Landsat Surface Reflectance, gridMET, and the National Land Cover Database.
}
