% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM.R
\name{viz_PRISM}
\alias{viz_PRISM}
\title{Visualize PRISM data from GEE}
\usage{
viz_PRISM(
  aoi,
  method = "Norm81m",
  param = "ppt",
  startDate = "1981-01-01",
  endDate = "2010-01-01",
  window = FALSE,
  w.low = NULL,
  w.high = NULL,
  c.low = 1,
  c.high = 12
)
}
\arguments{
\item{aoi}{A sf object.}

\item{method}{\code{character}. 'Norm81m'(default), 'AN81m', 'AN81d'.}

\item{param}{\code{character}. 'ppt' (default), tmin, tmax, tmean, tdmean (deg C); vpdmin, vpdmax (hPa)}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{window}{\code{logical} whether to mask out certain ranges}

\item{w.low}{\code{numeric} low value for mask, e.g. greater than 'w.low'}

\item{w.high}{\code{numeric} high value for mask, e.g. less than 'w.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}
}
\value{
A leaflet map with selected parameters
}
\description{
This function takes a lot of arguments but in the end is very efficient to view
PRISM data. At a minimum, just providing a sf object will generate the Norm81m grid
and print out numerous statistics, e.g. mean, max, min, stdDev and sum. Grids are generated using PRISM (Parameter-elevation Regressions on Independent Slopes Model) and
Google Earth Engine via the \link[rgee]{rgee-package}. This makes it fast to explore these image collections
in R alongside your data analysis.
}
\note{
It's important to note that you need a GEE account and also this function depends
on \link[rgee]{rgee-package}. The time parameter options 'AN81m' and 'AN81d' can take a long time to run unless the date range is
changed. So be aware of time when running these for
large areas, e.g. gt 50 sq.miles can take up to 30 seconds to process.
}
