% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar.R
\name{viz_Sent1}
\alias{viz_Sent1}
\title{Visualize Sentinel-1 SAR GRD: C-band}
\usage{
viz_Sent1()
}
\arguments{
\item{aoi}{A sf object}

\item{param}{A \code{character} indicating what band to visualize, e.g. 'NPOL', 'VV', 'VVsd'.}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{window}{\code{logical} whether to mask out certain ranges}

\item{w.low}{\code{numeric} low value for mask, e.g. greater than 'w.low'}

\item{w.high}{\code{numeric} high value for mask, e.g. less than 'w.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}

\item{scale}{\code{numeric} indicating in meters what to scale the statistics by, e.g. min/max.}
}
\value{

}
\description{
The Sentinel-1 mission provides data from a
dual-polarization C-band Synthetic Aperture Radar (SAR) instrument.
This collection includes the S1 Ground Range Detected (GRD) scenes,
processed using the Sentinel-1 Toolbox to generate a calibrated,
ortho-corrected product. The collection is updated daily. New assets are ingested within two days after they become available.
}
