% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{viz_Terrain}
\alias{viz_Terrain}
\title{Visualize DEMs}
\usage{
viz_Terrain(
  aoi,
  param = "slope",
  window = FALSE,
  w.low = NULL,
  w.high = NULL,
  stats = FALSE,
  az = 270,
  el = 45,
  scale = 10
)
}
\arguments{
\item{aoi}{A sf object}

\item{param}{\code{character} indicating terrain type, e.g. 'slope', 'TPI', 'sin', 'hillshade', 'complete'.}

\item{window}{\code{logical} whether to mask out certain ranges}

\item{w.low}{\code{numeric} low value for mask, e.g. greater than 'w.low'}

\item{w.high}{\code{numeric} high value for mask, e.g. less than 'w.high'}

\item{stats}{\code{logical} whether to compute stats or not, slope only.}

\item{az}{\code{numeric} The illumination azimuth in degrees from north. o}

\item{el}{\code{numeric} The illumination elevation in degrees. optional}

\item{scale}{\code{numeric} value indicating what to reduce the regions by, e.g. 10 (m) default.}
}
\value{

}
\description{
This function takes the USGS NED layer and generates a handfull of terrain visualizations. This is good for
getting a quick peak into an unknown area or just masking out certain ranges, e.g. slope > 45\%.
}
