% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_filters.R
\name{ee_month_filter}
\alias{ee_month_filter}
\alias{ee_month_filter.ee.imagecollection.ImageCollection}
\alias{ee_month_filter.exploreList}
\title{Filter by Month}
\usage{
ee_month_filter(imageCol, ...)

\method{ee_month_filter}{ee.imagecollection.ImageCollection}(imageCol, stat, months, ...)

\method{ee_month_filter}{exploreList}(imageCol, stat, ...)
}
\arguments{
\item{imageCol}{An earth engine ImageCollection}

\item{...}{extra args to pass on}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by,
e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'sd', 'first'.}

\item{months}{\code{numeric} vector, e.g. c(1,12).}
}
\description{
Filter by Month
}
