% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_filters.R
\name{ee_year_filter}
\alias{ee_year_filter}
\alias{ee_year_filter.ee.imagecollection.ImageCollection}
\alias{ee_year_filter.exploreList}
\title{Filter by Year}
\usage{
ee_year_filter(imageCol, stat, ...)

\method{ee_year_filter}{ee.imagecollection.ImageCollection}(imageCol, startDate, endDate, stat)

\method{ee_year_filter}{exploreList}(imageCol, stat)
}
\arguments{
\item{imageCol}{An earth engine ImageCollection}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by,
e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'sd', 'first'.}

\item{startDate}{\code{character} format date, e.g. "2018-10-23".}

\item{endDate}{\code{character} format date, e.g. "2018-10-23".}
}
\description{
Filter by Year
}
