% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_diff}
\alias{get_diff}
\title{Get Image Difference}
\usage{
get_diff(data, startDate2 = "2010-01-01", endDate2 = "2018-10-30", band = NULL)
}
\arguments{
\item{data}{A previously create get_* object}

\item{startDate2}{\code{character} format date, e.g. "1999-10-23". A second start date to use as the subtraction window.}

\item{endDate2}{\code{character} format date, e.g. "1999-10-23". A second end date to use as the subtraction window.}

\item{band}{A \code{character} indicating what bands/type to use when you have more than one. Can \strong{only} select one, e.g. 'NDVI'.}
}
\value{
A list of Earth Engine Objects and arguments.
}
\description{
This function allows the user to get differences by subtracting two images. Only works for one-band selections and
from different time series.
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

# get intitial image. Can by any get_*().

ld8 <- get_landsat(huc, method = 'ld8', startDate = '2014-01-01',
                  endDate = '2018-12-31', c.low = 6, c.high = 11)

# now subtract from new one

diff <- ld8 \%>\% get_diff(startDate2 = '2019-01-01', endDate2 = '2019-12-31', band = 'NDVI')

# now viz, rr, band, etc.

diff \%>\% viz(scale = 30)
}
}
