% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_landsat}
\alias{get_landsat}
\title{Get Landsat SR Products}
\usage{
get_landsat(
  aoi,
  method = "ld8",
  param = NULL,
  stat = "median",
  cloud_mask = TRUE,
  startDate = "1984-01-01",
  endDate = "2020-10-30",
  mask = FALSE,
  m.low = NULL,
  m.high = NULL,
  c.low = 1,
  c.high = 12
)
}
\arguments{
\item{aoi}{A sf object indicating the extent of the geom.}

\item{method}{A \code{character} indicating what method to use, e.g. 'ld8', 'ld7', 'ld5', 'ld4', 'ts', 'harm_ts'.}

\item{param}{A \code{character} indicating what band to visualize, e.g. 'Blue' or c('Green', 'Red', 'NIR') or NULL returns all bands.}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by, e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'stdDev', 'first'.}

\item{cloud_mask}{\code{logical} whether to mask out certain cloud artifacts. TRUE (default).}

\item{startDate}{\code{character} format date, e.g. "1999-10-23"}

\item{endDate}{\code{character} format date, e.g. "1999-10-23"}

\item{mask}{\code{logical} whether to mask out certain ranges}

\item{m.low}{\code{numeric} low value for mask, e.g. greater than 'm.low'}

\item{m.high}{\code{numeric} high value for mask, e.g. less than 'm.high'}

\item{c.low}{\code{numeric} lower month value for calendar range}

\item{c.high}{\code{numeric} higher month value for calendar range}
}
\value{
A list of Earth Engine Objects and arguments.
}
\description{
This function gets 30-m USGS Landsat 8, 7 , 5 and 4 (Surface Reflectance Tier 1 (SR))
satellite images from various bands. The image processing for time series can either use methods by Roy et al. (2016) (e.g. harmonization) so that all three Landsat
missions will be scaled for a harmonized time series analysis ('harm_ts') or you can just get time series without harmonizing ('ts'). All the methods will use
Zhu et al. (2015) masking function and filter image properties by cloud cover less than 50\% and Geometric RMSE Model less than 10.
}
\details{
The methods currently available (more to come):
\itemize{
\item  \strong{harm_ts}: Harmonizing Landsat Missions; 1984-01-01 - 2021-01-22.
\item  \strong{ts}: Combining Landsat Missions; 1984-01-01 - 2021-01-22.
\item \strong{ld8}: LANDSAT/LC08/C01/T1_SR; 2013-04-11 - 2021-01-22.
\item \strong{ld7}: LANDSAT/LC07/C01/T1_SR; 1999-01-01 - 2021-01-21.
\item \strong{ld5}: LANDSAT/LC05/C01/T1_SR; 1984-01-01 - 2012-05-05.
\item \strong{ld4}: LANDSAT/LC04/C01/T1_SR; 1982-08-22 - 1993-12-14.
}
The param (bands) currently available (more to come):
\itemize{
\item  \strong{Blue}, \strong{Green}, \strong{Red}, \strong{NIR},\strong{SWIR1},
\strong{SWIR2}, \strong{NDVI}, \strong{NDWI}, \strong{NBR}
}
}
\note{
If you want to use this with \href{https://github.com/davemlz/eemont}{#eemont} functions then make sure \code{cloud_mask} is set to \code{FALSE}.
Also, 'harm_ts' and 'ts' can take a long time to process in other function \link[exploreRGEE]{viz}, \link[exploreRGEE]{rr} and \link[exploreRGEE]{band}  so be aware!
}
\examples{
\dontrun{
# Load Libraries

library(rgee)
rgee::ee_intialize()
library(exploreRGEE)

# Bring in data
huc <- exploreRGEE::huc

ld8 <- get_landsat(huc, method = 'ld8', startDate = '2014-01-01',
                  endDate = '2018-12-31', c.low = 6, c.high = 11)


}
}
